Organise all the classes into packages
