+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.task.Deadline;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
