+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.task.ToDo;
+import duke.ui.Ui;
+
 public class ToDoCommand extends Command {
 
     public static final String COMMAND_WORD = "todo";
