+++ src/main/java/duke/exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class DukeException extends RuntimeException{
 
     public DukeException(String errorMessage) {
