+++ src/main/java/duke/exception/InvalidDateTimeException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class InvalidDateTimeException extends DukeException {
 
     public InvalidDateTimeException(String message) {
