+++ src/main/java/duke/exception/InvalidInputException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class InvalidInputException extends DukeException{
 
     public InvalidInputException() {
