+++ src/main/java/duke/exception/MissingDeadlineDescriptionException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class MissingDeadlineDescriptionException extends DukeException {
 
     public MissingDeadlineDescriptionException() {
