+++ src/main/java/duke/exception/MissingDescriptionException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class MissingDescriptionException extends DukeException {
 
     public MissingDescriptionException(String task) {
