+++ src/main/java/duke/exception/MissingEventDescriptionException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class MissingEventDescriptionException extends DukeException {
 
     public MissingEventDescriptionException() {
