+++ src/main/java/duke/exception/MissingTargetException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class MissingTargetException extends DukeException {
 
     public MissingTargetException(String command) {
